Wallet interface verifyMessage spec
