+++ src/Wallet/Wallet.spec.ts
@@ -1,0 +1,39 @@
+import { expect } from 'chai'
+import { MemoryKey } from '../KeyManager'
+import Wallet, { AddressType } from '.'
+
+describe('Wallet', () => {
+  describe('verifyMessage', () => {
+    it('returns true when verifying a correct signature for a message', () => {
+      const message = 'foobar'
+      const mnemonic = 'height bubble drama century ask online stage camp point loyal hip awesome'
+      const keypair = MemoryKey({ mnemonic, password: 'securepassword' })
+      const signatureAsHex = keypair.signMessage(AddressType.external, 0, message)
+
+      const verification = Wallet(keypair.publicAccount()).verifyMessage({
+        addressType: AddressType.external,
+        message,
+        signingIndex: 0,
+        signatureAsHex
+      })
+
+      expect(verification).to.eql(true)
+    })
+
+    it('returns false when verifying an incorrect message for a valid signature', () => {
+      const message = 'foobar'
+      const mnemonic = 'height bubble drama century ask online stage camp point loyal hip awesome'
+      const keypair = MemoryKey({ mnemonic, password: 'securepassword' })
+      const signatureAsHex = keypair.signMessage(AddressType.external, 0, message)
+
+      const verification = Wallet(keypair.publicAccount()).verifyMessage({
+        addressType: AddressType.external,
+        message: 'a different message',
+        signingIndex: 0,
+        signatureAsHex
+      })
+
+      expect(verification).to.eql(false)
+    })
+  })
+})
